requirement.txt
tensorboard
pandas
numpy
tqdm
torchsummary
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# pip install torch==1.8.1+cu101 torchvision==0.9.1+cu101 torchaudio==0.8.1 -f https://download.pytorch.org/whl/torch_stable.html
